+++ frontend/src/app/app-routing.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { HomeComponent } from './home/home.component';
 import { LoginComponent } from './login/login.component';
-import { ProfileComponent } from "./profile/profile.component";
-import { ProfileEditComponent} from "./profile-edit/profile-edit.component";
+import { ProfileShowComponent } from "./profile/profile-show/profile-show.component";
+import { ProfileEditComponent} from "./profile/profile-edit/profile-edit.component";
 import {PasswordEditComponent} from "./password-edit/password-edit.component";
 import {FacebookComponent} from "./facebook/facebook.component";
 
@@ -12,7 +12,7 @@ const routes: Routes = [
     { path: 'login', component: LoginComponent },
 
     { path: 'users', component: FacebookComponent },
-    { path: 'user/:id', component: ProfileComponent },
+    { path: 'user/:id', component: ProfileShowComponent },
     { path: 'user/:id/edit', component: ProfileEditComponent },
 
     { path: 'monprofil/editer/motdepasse', component: PasswordEditComponent }
