+++ frontend/src/app/profile/profile-edit/profile-edit.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {ApiService} from "../api.service";
+import {ApiService} from "../../api.service";
 import {ActivatedRoute, Router} from "@angular/router";
 
 @Component({
