+++ frontend/src/app/profile/profile-show/profile-show.component.html
@@ -22,13 +22,7 @@
             </div>
 
             <div class="card">
-                <div class="card-header">
-                    <div class="card-title">Informations personnelles</div>
-                </div>
                 <div class="card-body">
-                    <strong>Prénom : </strong> {{ user.first_name }} <br>
-                    <strong>Nom : </strong> {{ user.last_name }} <br>
-                    <br>
                     <strong>Parrain : </strong> {{ user.minesparent }} <br>
                     <br>
                     <strong>Email : </strong> {{ user.email }} <br>
