+++ frontend/src/app/profile/profile-show/profile-show.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { ProfileComponent } from './profile.component';
+import { ProfileComponent } from './profile-show.component';
 
 describe('ProfileComponent', () => {
   let component: ProfileComponent;
